# brainX_back/Intelligence-Service/src/main/resources/contracts/knowledge-intelligence.openapi.yaml
openapi: 3.0.3
info:
  title: BrainX API SSOT - Knowledge Intelligence
  version: 1.0.0-revised
  license:
    name: Proprietary
    url: https://brainx.p-e.kr
  description: 'BrainX 공개 및 내부 REST 계약 문서.


    개정 정책:

    - Public/client API는 현재 구현된 Ingestion 발행 헬퍼(/v1/publish-jobs)를 제외하고 /api/v1 아래에 둔다.

    - 서비스 간 동기 API는 /internal/v1 아래에 명시적으로 격리한다.

    - 중복된 회원 탈퇴 취소 엔드포인트는 제거했다.

    - /support/inquiries는 /support/tickets로 통합했다.

    - 토큰 사용량 기록은 AsyncAPI 이벤트 우선으로 처리하며 공개 REST command는 제공하지 않는다.

    - 발행 작업은 현재 클립보드용 콘텐츠를 포함해 200 OK를 반환한다.

    - operation 단위 소비자 메타데이터는 x-consumers로 표현한다.

    - operation 단위 이벤트 결합은 x-produces-events로 표현하고 AsyncAPI SSOT를 참조한다.'
servers:
- url: https://brainx.p-e.kr
  description: AWS dev API 게이트웨이. 로컬 개발은 Gateway-Service 기본 포트 8088을 사용한다.
- url: https://api.brainx.com
  description: 운영 API 게이트웨이
...
paths:
...
  /api/v1/ai/clusters:
    post:
      operationId: requestAiClusterJob
      summary: AI 클러스터링 요청
      tags:
      - Knowledge Intelligence
      x-producer-service: knowledge-intelligence
      x-visibility: public
      x-consumers:
      - id: web.graph
        type: frontend
        description: 2D/3D 그래프 및 마인드맵 UI
      x-produces-events:
      - ClusterJobRequested
      - TokenUsageRecordedRequested
      x-internal-sync-calls:
      - targetService: commerce-operations
        operationId: checkEntitlementsInternal
        reason: capability/quota must be decided synchronously before model call or job acceptance
        required: true
      x-async-boundary: true
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClusterJobCreateRequest'
      responses:
        '202':
          description: 성공
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiSuccessBase'
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      $ref: '#/components/schemas/ClusterJobData'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: 충돌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
      - bearerAuth: []
  /api/v1/ai/clusters/latest:
    get:
      operationId: getLatestAiClusterJob
      summary: 최신 AI 클러스터링 상태 조회
      tags:
      - Knowledge Intelligence
      x-producer-service: knowledge-intelligence
      x-visibility: public
      x-consumers:
      - id: web.graph
        type: frontend
        description: 2D/3D 그래프 및 마인드맵 UI
      x-produces-events: []
      x-internal-sync-calls: []
      x-async-boundary: false
      parameters:
      - name: documentGroupId
        in: query
        required: false
        schema:
          type: string
          default: default
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiSuccessBase'
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      $ref: '#/components/schemas/ClusterJobLatestData'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
      - bearerAuth: []
  /api/v1/ai/clusters/{clusterJobId}:
    get:
      operationId: getAiClusterJob
      summary: AI 클러스터링 결과 조회
      tags:
      - Knowledge Intelligence
      x-producer-service: knowledge-intelligence
      x-visibility: public
      x-consumers:
      - id: web.graph
        type: frontend
        description: 2D/3D 그래프 및 마인드맵 UI
      x-produces-events: []
      x-internal-sync-calls: []
      x-async-boundary: false
      parameters:
      - name: clusterJobId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiSuccessBase'
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      $ref: '#/components/schemas/ClusterJobData'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: 충돌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
      - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
      description: 명령 및 작업 요청에 권장한다. 같은 키로 재시도하면 같은 결과를 반환해야 한다.
  schemas:
    ApiSuccessBase:
      type: object
      properties:
        success:
          type: boolean
          enum:
          - true
        message:
          type: string
      required:
      - success
      - message
      additionalProperties: false
    ApiErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          enum:
          - false
        data:
          type: object
          nullable: true
        message:
          type: string
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            traceId:
              type: string
            details:
              type: object
              additionalProperties: true
              description: 기계가 읽을 수 있는 오류 상세 정보
          required:
          - code
          - message
          additionalProperties: true
      required:
      - success
      - message
      additionalProperties: false
    JobStatus:
      type: string
      enum:
      - PENDING
      - RUNNING
      - COMPLETED
      - FAILED
      - CANCELLED
...
    ClusterJobCreateRequest:
      type: object
      properties:
        scope:
          type: object
          additionalProperties: true
        algorithmOptions:
          type: object
          additionalProperties: true
      required:
      - scope
      additionalProperties: false
    ClusterJobData:
      type: object
      properties:
        clusterJobId:
          type: string
        documentGroupId:
          type: string
        status:
          $ref: '#/components/schemas/JobStatus'
        clusters:
          type: array
          items:
            type: object
            additionalProperties: true
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        failureMessage:
          type: string
          nullable: true
      required:
      - clusterJobId
      - documentGroupId
      - status
      - createdAt
      additionalProperties: false
    ClusterJobLatestState:
      type: string
      enum:
      - NO_SOURCE_NOTES
      - NOT_ANALYZED
      - FRESH
      - STALE
      - FAILED
    ClusterJobLatestData:
      type: object
      properties:
        documentGroupId:
          type: string
        searchableNoteCount:
          type: integer
          format: int32
          minimum: 0
        latestNoteUpdatedAt:
          type: string
          format: date-time
          nullable: true
        state:
          $ref: '#/components/schemas/ClusterJobLatestState'
        job:
          type: object
          nullable: true
          allOf:
          - $ref: '#/components/schemas/ClusterJobData'
      required:
      - documentGroupId
      - searchableNoteCount
      - state
      additionalProperties: false
...